% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oda_list_filters.R
\name{oda_list_filters}
\alias{oda_list_filters}
\title{List Available Filters for an ODA Resource}
\usage{
oda_list_filters(resource = NULL)
}
\arguments{
\item{resource}{A character string identifying the resource
(e.g., \code{"DSD_CRS@DF_CRS"}).}
}
\value{
A character vector of available filter names for the given resource.
}
\description{
List Available Filters for an ODA Resource
}
\examples{
oda_list_filters("DSD_CRS@DF_CRS")

oda_list_filters("DSD_DAC2@DF_DAC2A")

}
