% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oecdoda-package.R
\docType{package}
\name{oecdoda-package}
\alias{oecdoda}
\alias{oecdoda-package}
\title{oecdoda: Seamless Access to OECD Official Development Assistance (ODA) Data}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tidy-intelligence/r-oecdoda}
  \item \url{https://tidy-intelligence.github.io/r-oecdoda/}
  \item Report bugs at \url{https://github.com/tidy-intelligence/r-oecdoda/issues}
}

}
\author{
\strong{Maintainer}: Christoph Scheuch \email{christoph@tidy-intelligence.com} (\href{https://orcid.org/0009-0004-0423-6819}{ORCID}) [copyright holder]

}
\keyword{internal}
